requirements.txt
Pillow>=9.5.0
flask>=3.0.2
APScheduler==3.10.4 